Improve performance
This patch improves the performance of the `end_to_end()` test by about
a factor of 2. Using BTree-based collections for cases where we iterate
over the collection improves performance. Also, this patch avoids an
allocation in a hot spot, in `InputDataGraph::intersection()`.

This performance increase was observed with both optimization level 3
and "s" (because we build the WebAssembly version with opt-level "s"),
with LTO enabled in both cases.